Support Export Tool
This change adds the new support and support export commands to the pgo
cli. These commands export a snapshot of a postgrescluster when it is
invoked. The exported files are combined into a tar file in the defined
output directory.

The export collects the following:
- kubernetes server version
- list of nodes in the kubernetes cluster
- namespace yaml for the current namespace
- postgrescluster yaml
- API Resources related to a postgrescluster
- events in the current namespace
- postgres operator logs from the primary
- pod logs from every container
- output from the patronictl info and history commands

Kuttl tests are added to create a cluster and perform a support export