Create Command to View Backup Info
Adds a command to the pgo CLI to show the output of the 'pgbackrest
info' command. It allows for flags to specify both the repo name
and the output type. It assumes a user running pgo will have 'exec'
RBAC permissions to facilitate executing the necessary pgBackRest
command.

Issue: [sc-14640]