Delete PostgresCluster command
Adds a CLI command to delete a PostgresCluster. The command to
delete a cluster named 'hippo' would be as follows:

kubectl-pgo delete postgrescluster hippo

It takes one argument, the cluster's name and gives a warning that
this command is destructive and retention is based on PV policy.
For the command to continue, the user must provide confirmation.

Issue: [sc-14895]