PGO version command
Adds a command 'version' that returns the CLI and Operator versions.
The CLI version is set via a string constant.
The Operator version is retrieved from the 'app.kubernetes.io/version'
Label on the PostgresCluster CRD. If the CRD is not found, an error
is returned with the CLI version. If the relevant Label is missing,
a 'not found' message is returned with the CLI version.

Issue: [sc-15178]