Add monitoring Pod log capture and improve associated tests
This update adds the logs from the Grafana, Prometheus and
Alertmanager Pods to the support export archive. By default, the
support export command will look in the same namespace as the
given PostgresCluster. For cases where the monitoring Pods run in
a separate namespace, an override flag, --monitoring-namespace, is
provided. Note that monitoring logs are only gathered from a single
namespace at a time.

Additionally, this update adds associated tests for the monitoring
logs to include the Pods mentioned above and the Postgres exporter
sidecar container when defined in the PostgresCluster spec.

Issues:
[sc-17848]
[sc-17691]