Override Prometheus container command
This change overrides the Prometheus container's default command to
avoid a 'permission denied' error in some environments (e.g. Openshift).

This happens because the Kuttl test is defining a 'dummy' deployment
that is not configured for actual use. As such, overriding the command
does not impact the validity of the test.

Issue: [sc-17848]