Require a Postgres version when creating a PostgresCluster
This change adds a new flag 'pg-major-version' to the 'create
postgrescluster' command to allow a specific Postgres version to
be used for the PostgresCluster. The new flag is required and does
not default to a particular version. As before, the image used by
the cluster will be pulled from the corresponding 'RELATED_IMAGE'
value set in the environment.

KUTTL test updates are also included, along with a bump to Postgres
15 and a timeout adjustment.

Issue: PGO-446