Refactor NewPodExecutor function outside of loop.
Refactor error handling for finding PG and backrest logs. I refactored the error handling to use an OR condition to capture the state where no logs are available. This is because err and/or stderr may report 'no results' depending upon the way we call `ls` (i.e.  with/without piping the results to `head`),