Large postgres logs (#105)
Support for large Postgres files

Large Postgres files may pose a challenge for some systems. This change stores the file on disk, compares the filesize with the remote, and adds the local file to the tar. It has error handling to allow a support export to be generated when there are issues with gathering PG logs.

* fixes for lint errors

* Remove the cat logic for gathering PG logs. Only do streaming.

* Create a unique sudirectory in outputDir (matching the unique timestamp) so that local files are logically separate from other files. Adding messaging and error handling so a user understands.

* remove extra call to getRemoteFileSize

* refactored getRemoteFileSize to use bashCommand to get the filesize

* added copyFile function to cat a file contents to a local file

* use the copyFile function rather calling manually

* added example comment for fileSpecSrc

* added comment to copyFile

* rename ConvertBytes to convertBytes

---------

Co-authored-by: Philip Hurst <philip.hurst@crunchydata.com>