include kubectl describe output (#137)
* gather pg_settings

Query pg_settings to capture current Postgres settings directly from the DB.

* run kubectl describe commands on various resources

* removing cluster-info

cluster-info returns a lot of sensitive information and it out-of-scope for the support export.

* getting CRDs

Remove the `kubectl describe crds` and replace with getting the CRDs via the correct APIs. This brings CRDs into the support export output on the same level as pods, sts, etc.

* fix unnecessary use of fmt.Sprintf

* better gathering of PGAdmin resources

Treats PGAdmin resources separately since they are independent of a PostgresCluster. Gathers all PGAdmin resources in a namespace. This is usually one, but it could be several.

* fix describe for clusterrole and clusterrolebinding on openshift

* refactor describe clusterrole and clusterrolebinding

This makes it more clear that 'postgresoperator' is a special case for certain installers.

* refactor to account for labels on some installers

Some installers don't label the CRD with app.kubernetes.io/name=pgo. This refactor filters all the CRDs for a name containing "postgres-operator.crunchydata.com"