Migrating PGO CLI kuttl tests to chainsaw (#138)
* Kyverno policies

We can use Kyverno policies to facilitate testing with different image repos.

* initial set of templates

These are helpful templates to use when testing the PGO CLI

* default config and values YAML files

* Chainsaw test for PGO CLI backup

* Chainsaw test for PGO CLI backup

* Chainsaw tests for PGO CLI create

* Chainsaw tests for PGO CLI delete

* Chainsaw test for PGO CLI show

* Chainsaw tests for PGO CLI version

* Add Chainsaw README

Instructions for how to run the tests

* remove skipDelete

The chainsaw-show Test should remove resources after it completes.

* Chainsaw test for PGO CLI start-stop

* Chainsaw test for PGO CLI restore

* Update chainsaw tests with helpful sleeps

* Update template with better check for backup complete

* added newlines at the end

* renamed steps in version test

* refactored backupSelector

The way we defined the selector for backups was quite terse. This refactor simplifies the logic.

* Chainsaw test for PGO CLI support export